Separate Tensor and TensorCanonicalizer in different headers
