+++ SeQuant/core/tensor.cpp
@@ -6,6 +6,7 @@
 #include <SeQuant/core/abstract_tensor.hpp>
 #include <SeQuant/core/expr.hpp>
 #include <SeQuant/core/index.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
 
 namespace sequant {
