+++ examples/eval/btas/main.cpp
@@ -6,6 +6,7 @@
 
 #include <btas/btas.h>
 #include <SeQuant/core/op.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
 #include <SeQuant/domain/mbpt/context.hpp>
 #include <SeQuant/domain/mbpt/convention.hpp>
