+++ examples/osstcc/osstcc.cpp
@@ -1,4 +1,5 @@
 #include <SeQuant/core/op.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
 #include <SeQuant/core/timer.hpp>
 #include <SeQuant/domain/mbpt/context.hpp>
 #include <SeQuant/domain/mbpt/convention.hpp>
