+++ python/src/sequant/mbpt.h
@@ -7,6 +7,7 @@
 #include <SeQuant/core/abstract_tensor.hpp>
 #include <SeQuant/core/context.hpp>
 #include <SeQuant/core/expr.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
 
 #include "python.h"
