+++ tests/unit/test_canonicalize.cpp
@@ -8,6 +8,7 @@
 #include <SeQuant/core/index.hpp>
 #include <SeQuant/core/latex.hpp>
 #include <SeQuant/core/rational.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
 
 #include <memory>
 #include <string>
