+++ tests/unit/test_eval_expr.cpp
@@ -9,6 +9,7 @@
 #include <SeQuant/core/container.hpp>
 #include <SeQuant/core/expr.hpp>
 #include <SeQuant/core/index.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
 
 #include <string_view>
 #include <initializer_list>
