+++ tests/unit/test_tensor_network.cpp
@@ -17,6 +17,7 @@
 #include <SeQuant/core/expr.hpp>
 #include <SeQuant/core/hash.hpp>
 #include <SeQuant/core/index.hpp>
+#include <SeQuant/core/tensor_canonicalizer.hpp>
 
 #include <iostream>
 #include <cstdlib>
